requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
uvicorn==0.11.5
numpy==1.18.4
pathlib==1.0.1
python-multipart==0.0.5
starlette==0.13.4
pillow==7.1.2
fastai==1.0.61
torch==1.5.0
torchvision==0.6.0

